requirements.txt
python-garminconnect==0.1.21
pandas==2.1.0
numpy==1.24.3
streamlit==1.27.0
matplotlib==3.7.3
plotly==5.17.0
pytest==7.4.2
pytest-mock==3.11.1
sqlalchemy==2.0.21
python-dotenv==1.0.0
requests==2.31.0